Implement q_insert_head and q_insert_tail
First allocate a pointer to element_t struct and check if any of the pointers
to structs are NULL.

Then use list_add or list_add_tail implemented in list.h to make new list_head
linked to head.

Use strdup to copy input char list and assign it to new_element->value.